bin/requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile ./bin/requirements.in
#
amqp==5.1.1
    # via kombu
async-timeout==4.0.2
    # via redis
billiard==3.6.4.0
    # via celery
celery[redis]==5.2.7
    # via -r ./bin/requirements.in
certifi==2022.6.15
    # via requests
charset-normalizer==2.1.0
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-default-group
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   shot-scraper
    #   typer
click-default-group==1.2.2
    # via shot-scraper
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
commonmark==0.9.1
    # via rich
dateutils==0.6.12
    # via -r ./bin/requirements.in
deprecated==1.2.13
    # via redis
greenlet==1.1.2
    # via playwright
idna==3.3
    # via requests
inflection==0.5.1
    # via -r ./bin/requirements.in
kombu==5.2.4
    # via celery
packaging==21.3
    # via redis
playwright==1.24.0
    # via
    #   -r ./bin/requirements.in
    #   shot-scraper
prompt-toolkit==3.0.30
    # via click-repl
pydantic==1.9.1
    # via -r ./bin/requirements.in
pyee==8.1.0
    # via playwright
pygments==2.12.0
    # via rich
pyparsing==3.0.9
    # via packaging
python-dateutil==2.8.2
    # via dateutils
python-frontmatter==1.0.0
    # via -r ./bin/requirements.in
python-slugify==6.1.2
    # via -r ./bin/requirements.in
pytz==2022.1
    # via
    #   celery
    #   dateutils
pyyaml==6.0
    # via
    #   python-frontmatter
    #   shot-scraper
redis==4.3.4
    # via celery
requests==2.28.1
    # via -r ./bin/requirements.in
rich==12.5.1
    # via -r ./bin/requirements.in
shot-scraper==0.14.1
    # via -r ./bin/requirements.in
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
text-unidecode==1.3
    # via python-slugify
typer==0.6.1
    # via -r ./bin/requirements.in
typing-extensions==4.3.0
    # via pydantic
urllib3==1.26.10
    # via requests
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
websockets==10.1
    # via playwright
wrapt==1.14.1
    # via deprecated
